Rename channels.rs to benchmarking.rs
